uni-data/requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
markdownify==0.12.1
mysql-connector-python==8.4.0
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.2
six==1.16.0
soupsieve==2.5
tzdata==2024.1
urllib3==2.2.1
SQLAlchemy==2.0.29